refactor: fix runner types
