+++ unicon_backend/evaluator/tasks/base.py
@@ -4,6 +4,8 @@ from typing import Any, Generic, TypeVar
 
 from pydantic import BaseModel
 
+from unicon_backend.evaluator.tasks.programming.runner import TaskEvalStatus
+
 TaskUserInput = TypeVar("TaskUserInput")
 TaskExpectedAnswer = TypeVar("TaskExpectedAnswer")
 TaskResult = TypeVar("TaskResult")
@@ -14,13 +16,6 @@ class TaskType(str, Enum):
     MULTIPLE_RESPONSE = "MULTIPLE_RESPONSE_TASK"
     SHORT_ANSWER = "SHORT_ANSWER_TASK"
     PROGRAMMING = "PROGRAMMING_TASK"
-
-
-class TaskEvalStatus(str, Enum):
-    SUCCESS = "SUCCESS"
-    PENDING = "PENDING"
-    SKIPPED = "SKIPPED"
-    FAILED = "FAILED"
 
 
 class TaskEvalResult(BaseModel, Generic[TaskResult]):
